requirements_deploy.txt
# Lightweight requirements for Cloud Run deployment (ureca)
# NOTE: crewai is NOT included - a lightweight stub is created in the Dockerfile
streamlit>=1.40.0
python-docx>=1.1.0
python-dotenv>=1.0.0
Pillow>=10.4.0
pandas>=2.2.0
jinja2>=3.1.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
